pyproject.toml
[project]
name = "luma-portal"
version = "0.1.0"
description = "Event check-in portal with badge printing for Luma.com events"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiosqlite>=0.20.0",
    "thefuzz[speedup]>=0.22.0",
    "pillow>=11.0.0",
    "brother-ql>=0.9.4",
    "httpx>=0.28.0",
    "qrcode[pil]>=8.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
